Create OpenAPI specification (#72)
* Add OpenAPI specification that documents the backend endpoints.
* Update backend to use code generated from the OpenAPI specification.
* Update frontend to use code generated from the OpenAPI specification.

+++ backend/src/main/java/ca/metricalsky/winston/entity/view/VideoCountView.java
@@ -4,5 +4,5 @@ public interface VideoCountView {
 
     String getChannelId();
 
-    Long getVideos();
+    Integer getVideos();
 }
